music_tg_bot/requirements.txt
aiogram==3.4.1
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg[binary]==3.1.18
alembic==1.13.1
redis==5.0.1
rq==1.15.1
httpx==0.27.0
PyYAML==6.0.1
